[Feature #157262513]write a manage to run migration script and put all endpoints in a separate blueprint
