+++ app/auth/__init__.py
@@ -3,3 +3,4 @@ from flask import Blueprint
 auth = Blueprint('auth', __name__)
 
 from . import views
+
