+++ manage.py
@@ -1,0 +1,16 @@
+import os
+from flask_script import Manager
+from flask_migrate import Migrate, MigrateCommand
+
+# local imports
+from app import db, create_app
+from app import models
+
+app = create_app(config_name=os.getenv('APP_SETTINGS'))
+migrate = Migrate(app, db)
+manager = Manager(app)
+
+manager.add_command('db', MigrateCommand)
+
+if __name__ == '__main__':
+    manager.run()
